Move bundle selection types and EreportFilters to nexus-types
Move BundleDataCategory, BundleData, BundleDataSelection, and SledSelection
to nexus/types/src/support_bundle.rs. Move EreportFilters to
nexus/types/src/fm/ereport.rs (where it belongs alongside other ereport
types). Update all consumers to import directly.

Pure refactor, no behavior change.
